v0.3.1 Beta
Added logic to handle JRE and JDK.
Added uninstaller for Adobe Shockwave for when Shockwave is in the list
of required updates.
Temporary fix: Application now runs "Programs and Features" when JRE or
JDK is in the list of required updates so the user can more easily
manually uninstall the previous version.